requirements.txt
fabric==1.0.0
python-memcached
pymongo
djangotoolbox
django-mongodb-engine